rename generator and add simple conda one
