+++ generators/ueber-conda/templates/processed/withinEnv
@@ -1,0 +1,5 @@
+#!/usr/bin/env bash
+set -e
+source activate <%-env%>
+( exec $@ )
+source deactivate
