+++ generators/ueber-conda/templates/processed/withinEnv.cmd
@@ -1,0 +1,5 @@
+@echo off
+
+activate <%-env->
+%*
+deactivate
