+++ generators/ueber-vagrant/templates/plain/Vagrantfile_provision.sh
