+++ generators/ueber-vagrant/templates/processed/Vagrantfile
