+++ generators/ueber-virtualenv/templates/processed/setupVirtualenv.sh
